ei: reset w h x y when calculating screen size
These values were not reset each time the screen shape was update, so if
the screen shrink the values will be invalid and lead to borders not
triggered.
Also x and y need to be initialised at their maximum values or else they
will be always 0 since only min are called on them in the calculation
process.